refactor examples spec
