+++ package.json
@@ -86,7 +86,7 @@
     "chalk": "3.0.0",
     "cheerio": "1.0.0-rc.3",
     "common-tags": "1.8.0",
-    "cypress": "6.0.1",
+    "cypress": "6.1.0",
     "dependency-check": "3.4.1",
     "deps-ok": "1.4.1",
     "eslint": "6.1.0",
